+++ packages/pytest-simcore/src/pytest_simcore/helpers/catalog_data_fakers.py
